implemented 7edo easter egg
added in an easter egg where if the C button is held down when the module is powered on, the scale bank loaded is the 7Edo microtonal scale, which is a scale with 7 equal temperment tones per octave. The left column of 7 buttons act as the toggles for the scale, the right column do nothing. The C# button flashes when the easter egg mode is on. To load the normal western scale, the module must be restarted